Separate HID for platforms
